+++ docs/examples/guide/reactivity/examples/computed01/main.rs
@@ -89,9 +89,24 @@ impl TextualApp for ComputedApp {
             .with_child(
                 
                     Horizontal::new()
-                        .with_child(Input::new().with_placeholder("Enter red 0-255").id("red"))
-                        .with_child(Input::new().with_placeholder("Enter green 0-255").id("green"))
-                        .with_child(Input::new().with_placeholder("Enter blue 0-255").id("blue"))
+                        .with_child(
+                            Input::new()
+                                .with_value("0")
+                                .with_placeholder("Enter red 0-255")
+                                .id("red"),
+                        )
+                        .with_child(
+                            Input::new()
+                                .with_value("0")
+                                .with_placeholder("Enter green 0-255")
+                                .id("green"),
+                        )
+                        .with_child(
+                            Input::new()
+                                .with_value("0")
+                                .with_placeholder("Enter blue 0-255")
+                                .id("blue"),
+                        )
                 .id("color-inputs"),
             )
             .with_child(Static::new("").id("color"))